refactor: updateStatus의 반환타입을 Long에서 Integer로 변경

+++ src/main/java/org/example/ticketing/domain/reservation/repository/ReservationRepository.java
@@ -8,7 +8,7 @@ import reactor.core.publisher.Mono;
 
 public interface ReservationRepository extends ReactiveCrudRepository<Reservation, Long> {
   @Query("UPDATE reservation SET status = :status WHERE id = :id")
-  Mono<Long> updateStatus(Long id, Status status);
+  Mono<Integer> updateStatus(Long id, Status status);
 
   @Query("UPDATE reservation SET payment_id = :paymentId, status = :status WHERE id = :id")
   Mono<Integer> updatePaymentAndStatus(Long id, Long paymentId, Status status);
